+++ src/AggregateFunctions/examples/quantile-t-digest.cpp
@@ -1,6 +1,7 @@
 #include <AggregateFunctions/QuantileTDigest.h>
 #include <IO/WriteBufferFromString.h>
 #include <IO/ReadBufferFromString.h>
+#include <iostream>
 
 int main(int, char **)
 {
