+++ src/Columns/tests/gtest_column_vector.cpp
@@ -4,6 +4,7 @@
 #include <vector>
 #include <Columns/ColumnsNumber.h>
 #include <Common/randomSeed.h>
+#include <Common/thread_local_rng.h>
 #include <gtest/gtest.h>
 
 using namespace DB;
