+++ src/Common/Arena.h
@@ -8,6 +8,7 @@
 #include <Common/Allocator.h>
 #include <Common/ProfileEvents.h>
 #include <Common/memcpySmall.h>
+#include <base/getPageSize.h>
 
 #if __has_include(<sanitizer/asan_interface.h>) && defined(ADDRESS_SANITIZER)
 #   include <sanitizer/asan_interface.h>
