+++ src/Common/FiberStack.h
@@ -13,6 +13,11 @@
 #include <valgrind/valgrind.h>
 #endif
 
+/// Required for older Darwin builds, that lack definition of MAP_ANONYMOUS
+#ifndef MAP_ANONYMOUS
+#define MAP_ANONYMOUS MAP_ANON
+#endif
+
 namespace DB::ErrorCodes
 {
     extern const int CANNOT_ALLOCATE_MEMORY;
