+++ src/Common/OpenTelemetryTraceContext.cpp
@@ -2,6 +2,7 @@
 
 #include <random>
 #include <base/getThreadId.h>
+#include <Common/thread_local_rng.h>
 #include <Common/Exception.h>
 #include <base/hex.h>
 #include <Core/Settings.h>
