+++ src/IO/MMapReadBufferFromFileWithCache.cpp
@@ -1,4 +1,5 @@
 #include <IO/MMapReadBufferFromFileWithCache.h>
+#include <base/getPageSize.h>
 
 
 namespace DB
