+++ src/Interpreters/Cache/WriteBufferToFileSegment.cpp
@@ -7,6 +7,7 @@
 #include <base/scope_guard.h>
 
 #include <Common/logger_useful.h>
+#include <Common/formatReadable.h>
 
 namespace DB
 {
