+++ src/Processors/Merges/MergingSortedTransform.cpp
@@ -3,6 +3,7 @@
 #include <IO/WriteBuffer.h>
 
 #include <Common/logger_useful.h>
+#include <Common/formatReadable.h>
 
 namespace DB
 {
