+++ src/Storages/StorageS3Settings.cpp
@@ -6,6 +6,7 @@
 #include <Common/NamedCollections/NamedCollections.h>
 #include <Common/Exception.h>
 #include <Common/Throttler.h>
+#include <Common/formatReadable.h>
 #include <Interpreters/Context.h>
 #include <boost/algorithm/string/predicate.hpp>
